refactor : 리뷰 서비스 레이어, 컨트롤러 -> 하나의 클래스로 병합
