+++ src/test/java/app/bottlenote/review/service/ReviewServiceTest.java
@@ -34,6 +34,7 @@ import app.bottlenote.user.repository.UserCommandRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -68,134 +69,151 @@ class ReviewServiceTest {
 	private final Long reviewId = 1L;
 	private final Long userId = 1L;
 
-	@Test
-	@DisplayName("리뷰를 등록할 수 있다.")
-	void review_create_success() {
-		//given
+	@Nested
+	@DisplayName("리뷰 등록 서비스 레이어 테스트")
+	class ReviewCreateServiceTest {
 
-		//when
-		when(alcoholQueryRepository.findById(anyLong()))
-			.thenReturn(Optional.of(alcohol));
+		@Test
+		@DisplayName("리뷰를 등록할 수 있다.")
+		void review_create_success() {
+			//given
 
-		when(userCommandRepository.findById(anyLong()))
-			.thenReturn(Optional.of(user));
+			//when
+			when(alcoholQueryRepository.findById(anyLong()))
+				.thenReturn(Optional.of(alcohol));
 
-		when(reviewRepository.save(any(Review.class)))
-			.thenReturn(review);
+			when(userCommandRepository.findById(anyLong()))
+				.thenReturn(Optional.of(user));
 
-		ReviewCreateResponse response = reviewService.createReview(reviewCreateRequest, 1L);
+			when(reviewRepository.save(any(Review.class)))
+				.thenReturn(review);
 
-		verify(reviewImageSupport, times(1)).saveImages(anyList(), any());
-		verify(reviewTastingTagSupport, times(1)).saveReviewTastingTag(anyList(), any());
+			ReviewCreateResponse response = reviewService.createReview(reviewCreateRequest, 1L);
 
-		assertEquals(response.getId(), review.getId());
+			verify(reviewImageSupport, times(1)).saveImages(anyList(), any());
+			verify(reviewTastingTagSupport, times(1)).saveReviewTastingTag(anyList(), any());
+
+			assertEquals(response.getId(), review.getId());
+		}
+
+		@Test
+		@DisplayName("Alcohol이 존재하지 않을 때 AlcoholException이 발생해야 한다.")
+		void review_create_fail_when_alcohol_is_null() {
+			// given
+			when(alcoholQueryRepository.findById(anyLong())).thenReturn(Optional.empty());
+
+			// when, then
+			assertThrows(AlcoholException.class, () -> reviewService.createReview(reviewCreateRequest, 1L));
+		}
+
+		@Test
+		@DisplayName("유저가 존재하지 않을 때 UserNotFoundException 발생해야 한다.")
+		void review_create_fail_when_user_is_null() {
+			// given
+			when(alcoholQueryRepository.findById(anyLong())).thenReturn(Optional.of(alcohol));
+			when(userCommandRepository.findById(anyLong())).thenReturn(Optional.empty());
+
+			// when, then
+			assertThrows(UserException.class, () -> reviewService.createReview(reviewCreateRequest, 1L));
+		}
 	}
 
-	@Test
-	@DisplayName("Alcohol이 존재하지 않을 때 AlcoholException이 발생해야 한다.")
-	void review_create_fail_when_alcohol_is_null() {
-		// given
-		when(alcoholQueryRepository.findById(anyLong())).thenReturn(Optional.empty());
+	@Nested
+	@DisplayName("리뷰 조회 서비스 레이어 테스트")
+	class ReviewReadServiceTest {
 
-		// when, then
-		assertThrows(AlcoholException.class, () -> reviewService.createReview(reviewCreateRequest, 1L));
+		@Test
+		@DisplayName("리뷰를 조회할 수 있다.")
+		void testReviewRead() {
+			//given
+
+			//when
+			when(reviewRepository.getReviews(
+				anyLong(), any(PageableRequest.class), anyLong()))
+				.thenReturn(response);
+			PageResponse<ReviewListResponse> actualResponse = reviewService.getReviews(1L, request, 1L);
+
+			//then
+			assertThat(response.content()).isEqualTo(actualResponse.content());
+			assertThat(response.cursorPageable()).isEqualTo(actualResponse.cursorPageable());
+			verify(reviewRepository).getReviews(anyLong(), any(PageableRequest.class), anyLong());
+		}
+
+		@Test
+		@DisplayName("내가 작성한 리뷰를 조회할 수 있다.")
+		void testMyReviewRead() {
+			//given
+			Long userId = 1L;
+
+			//when
+			when(reviewRepository.getReviewsByMe(anyLong(), any(PageableRequest.class), anyLong()))
+				.thenReturn(response);
+
+			PageResponse<ReviewListResponse> actualResponse = reviewService.getMyReviews(1L, request, userId);
+
+			//then
+			assertThat(response.content()).isEqualTo(actualResponse.content());
+			assertThat(response.cursorPageable()).isEqualTo(actualResponse.cursorPageable());
+			verify(reviewRepository).getReviewsByMe(anyLong(), any(PageableRequest.class), anyLong());
+		}
 	}
 
-	@Test
-	@DisplayName("유저가 존재하지 않을 때 UserNotFoundException 발생해야 한다.")
-	void review_create_fail_when_user_is_null() {
-		// given
-		when(alcoholQueryRepository.findById(anyLong())).thenReturn(Optional.of(alcohol));
-		when(userCommandRepository.findById(anyLong())).thenReturn(Optional.empty());
+	@Nested
+	@DisplayName("리뷰 수정 서비스 레이어 테스트")
+	class ReviewModifyServiceTest {
 
-		// when, then
-		assertThrows(UserException.class, () -> reviewService.createReview(reviewCreateRequest, 1L));
+		@Test
+		@DisplayName("리뷰를 수정할 수 있다.")
+		void modify_review_success_when_without_tasting_tag() {
+
+			//when
+			when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenReturn(Optional.of(review));
+
+			reviewService.modifyReview(reviewModifyRequest, 1L, 1L);
+
+			verify(reviewImageSupport, times(1)).updateImages(anyList(), any());
+			verify(reviewTastingTagSupport, times(1)).updateReviewTastingTags(anyList(), any());
+			verify(reviewTastingTagSupport, never()).saveReviewTastingTag(any(), any());
+		}
+
+		@Test
+		@DisplayName("존재하지 않는 리뷰는 수정할 수 없다.")
+		void modify_review_fail_when_review_id_is_not_invalid() {
+			when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenThrow(
+				new ReviewException(ReviewExceptionCode.REVIEW_NOT_FOUND)
+			);
+
+			assertThrows(ReviewException.class, () -> reviewService.modifyReview(reviewModifyRequest, 100L, 1L));
+		}
 	}
 
+	@Nested
+	@DisplayName("리뷰 삭제 서비스 레이어 테스트")
+	class ReviewDeleteServiceTest {
 
-	@Test
-	@DisplayName("리뷰를 조회할 수 있다.")
-	void testReviewRead() {
-		//given
+		@Test
+		@DisplayName("리뷰를 삭제할 수 있다.")
+		void delete_review_success() {
 
-		//when
-		when(reviewRepository.getReviews(
-			anyLong(), any(PageableRequest.class), anyLong()))
-			.thenReturn(response);
-		PageResponse<ReviewListResponse> actualResponse = reviewService.getReviews(1L, request, 1L);
+			//given
+			when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenReturn(Optional.of(review));
 
-		//then
-		assertThat(response.content()).isEqualTo(actualResponse.content());
-		assertThat(response.cursorPageable()).isEqualTo(actualResponse.cursorPageable());
-		verify(reviewRepository).getReviews(anyLong(), any(PageableRequest.class), anyLong());
+			//when
+			ReviewResultResponse reviewResultResponse = reviewService.deleteReview(reviewId, userId);
+
+			//then
+			Assertions.assertEquals(DELETE_SUCCESS, reviewResultResponse.codeMessage());
+		}
+
+		@Test
+		@DisplayName("존재하지 않는 리뷰는 삭제할 수 없다.")
+		void delete_fail_when_review_is_not_exist() {
+
+			//when
+			when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenThrow(new ReviewException(REVIEW_NOT_FOUND));
+
+			//then
+			assertThrows(ReviewException.class, () -> reviewService.deleteReview(reviewId, userId));
+		}
 	}
-
-	@Test
-	@DisplayName("내가 작성한 리뷰를 조회할 수 있다.")
-	void testMyReviewRead() {
-		//given
-		Long userId = 1L;
-
-		//when
-		when(reviewRepository.getReviewsByMe(anyLong(), any(PageableRequest.class), anyLong()))
-			.thenReturn(response);
-
-		PageResponse<ReviewListResponse> actualResponse = reviewService.getMyReviews(1L, request, userId);
-
-		//then
-		assertThat(response.content()).isEqualTo(actualResponse.content());
-		assertThat(response.cursorPageable()).isEqualTo(actualResponse.cursorPageable());
-		verify(reviewRepository).getReviewsByMe(anyLong(), any(PageableRequest.class), anyLong());
-	}
-
-
-	@Test
-	@DisplayName("리뷰를 수정할 수 있다.")
-	void modify_review_success_when_without_tasting_tag() {
-
-		//when
-		when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenReturn(Optional.of(review));
-
-		reviewService.modifyReview(reviewModifyRequest, 1L, 1L);
-
-		verify(reviewImageSupport, times(1)).updateImages(anyList(), any());
-		verify(reviewTastingTagSupport, times(1)).updateReviewTastingTags(anyList(), any());
-		verify(reviewTastingTagSupport, never()).saveReviewTastingTag(any(), any());
-	}
-
-	@Test
-	@DisplayName("존재하지 않는 리뷰는 수정할 수 없다.")
-	void modify_review_fail_when_review_id_is_not_invalid() {
-		when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenThrow(
-			new ReviewException(ReviewExceptionCode.REVIEW_NOT_FOUND)
-		);
-
-		assertThrows(ReviewException.class, () -> reviewService.modifyReview(reviewModifyRequest, 100L, 1L));
-	}
-
-	@Test
-	@DisplayName("리뷰를 삭제할 수 있다.")
-	void delete_review_success() {
-
-		//given
-		when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenReturn(Optional.of(review));
-
-		//when
-		ReviewResultResponse reviewResultResponse = reviewService.deleteReview(reviewId, userId);
-
-		//then
-		Assertions.assertEquals(DELETE_SUCCESS, reviewResultResponse.codeMessage());
-	}
-
-	@Test
-	@DisplayName("존재하지 않는 리뷰는 삭제할 수 없다.")
-	void delete_fail_when_review_is_not_exist() {
-
-		//when
-		when(reviewRepository.findByIdAndUserId(anyLong(), anyLong())).thenThrow(new ReviewException(REVIEW_NOT_FOUND));
-
-		//then
-		assertThrows(ReviewException.class, () -> reviewService.deleteReview(reviewId, userId));
-	}
-
 }
